requirements.txt
annotated-types==0.7.0
anthropic==0.35.0
anyio==4.6.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jiter==0.5.0
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
